Support configuration, processing and validation of future returning signatures to declare asynchronous operations.
The ModuleGen annotation allows to override the callback signature model to a future returning signature model. When future is used, asynchronous callbacks become illegal and will fail the compilation. Future returning methods also have the future method kind and they can be distinguished by the code generator using a flag that indicates the module configuration, so generators can adapt the produced code to the actual Java method to call.

fixes #331