Introduce mechanism for specifying protobuf field numbers
Previously, field numbers were always assigned automatically during protobuf
schema generation, based on declaration order of properties in a data class.
It is very easy to modify the data class in a way that changes field numbers
unintentionally, which introduces a breaking change to the protobuf schema.

This commit introduces the annotation `@ProtobufField` as a mechanism for
explicitly providing field numbers, as well as a `FieldNumberStrategy` enum
(used in `@ProtobufGen.fieldNumberStrategy`) to select how field numbers
should be assigned. There's a manual strategy and two automatic strategies.

Further, `@ProtobufGen.reservedFieldNumbers` and `reservedFieldNames` allow
specifying reserved field numbers / names, similarly to the protobuf language
itself. Reserved field numbers and names are verified at compile time.

The fields in the protobuf schema are generated in the field number order,
not in the property declaration order. The serialization/deserialization
code in generated converters also uses field number order. This is to make
sure that a Vert.x-generated and `protoc`-generated serializers produce
outputs that are not just mutually _compatible_, but also binary _identical_.