JPMS support.
vertx-codegen-api declares its own module descriptor.

vertx-codegen-processor and vertx-codegen-json instead rely on moditect to generate descriptor for the sake of preserving the execution of codegen processor tests in IDE. The codegen processor is not intended to be modular, however vertx-codegen-api and vertx-codegen-json are actually consumed by from java modules (e.g io.vertx.core module).

the absence of an module-info descriptor instructs the IDE to run tests in classpath mode which is exactly what we need, hence the use of moditect for such trivial case.