Add support for java.time.Duration as part of the JSON generator.
Motivation:

The pattern (number, time unit) is often used to express a duration in Vert.x data objects, such as a timeout.

The type java.time.Duration is a good alternative and can be used to express the same property in a better fashion.

Changes:

The JSON generator handles now java.time.Duration and converts it back and forth a number in milliseconds.